unified buttons logic add/save delete/cancel

+++ hooks/useHandleButtons.ts
@@ -9,24 +9,35 @@ export default function useHandleButtons() {
   const date = UpdateDate();
   const color = '#bde0fe';
 
-  const handleSave = async (title: string | undefined, content: string | undefined) => {
+  const handleAdd = async (id: number, title: string, content: string) => {
     try {
-      await db.insert(notesTable).values([
-        {
-          title: title,
-          content: content,
-          color: color,
-          updatedDate: date,
-        },
-      ]);
+      if (id) {
+        await db
+          .update(notesTable)
+          .set({ title: title, content: content, color: color, updatedDate: date })
+          .where(eq(notesTable.id, id));
+      } else {
+        await db.insert(notesTable).values([
+          {
+            title: title,
+            content: content,
+            color: color,
+            updatedDate: date,
+          },
+        ]);
+      }
+
       router.back();
     } catch (error) {
       console.log('Error al agregar la nota.', error);
     }
   };
+
   const handleDelete = async (id: number) => {
     try {
-      await db.delete(notesTable).where(eq(notesTable.id, id));
+      if (id) {
+        await db.delete(notesTable).where(eq(notesTable.id, id));
+      }
       if (router.canGoBack()) {
         router.back();
       }
@@ -43,5 +54,5 @@ export default function useHandleButtons() {
     router.push({ pathname: '/notes/[id]', params: { id } });
   };
 
-  return { handleSave, handleDelete, handleNavigate };
+  return { handleAdd, handleDelete, handleNavigate };
 }
